:recycle: Improve structure for ProjectConfig
