+++ commands/command-new/src/commands/new.stage.ts
@@ -34,7 +34,7 @@ export type NewStageEvents = 'before.new:stage' | 'new:stage' | 'after.new:stage
 
 export class NewStage extends PluginCommand<NewStageEvents> {
   static id = 'new:stage';
-  static description = 'Creates a new stage.';
+  static description = 'Creates a new stage';
   static examples: string[] = [];
   // Defines arguments that can be passed to the command.
   static args = [
