+++ core/src/Config.ts
@@ -2,19 +2,18 @@ import _get from 'lodash.get';
 import _merge from 'lodash.merge';
 import _mergeWith from 'lodash.mergewith';
 import { join as joinPaths } from 'path';
-
 import { JovoCliError } from './JovoCliError';
 import { JovoCliPlugin } from './JovoCliPlugin';
 import { Log } from './Logger';
+import { ProjectConfig } from './ProjectConfig';
 import { mergeArrayCustomizer } from './utilities';
-import { ProjectConfigFile } from './interfaces';
 
 export class Config {
   private static instance?: Config;
-  private readonly config: ProjectConfigFile;
+  private readonly config: ProjectConfig;
 
   constructor(private projectPath: string, private stage?: string) {
-    const configContent: ProjectConfigFile = this.getContent();
+    const configContent: ProjectConfig = this.getContent();
     if (!stage) {
       this.stage = _get(configContent, 'defaultStage');
     }
@@ -38,11 +37,11 @@ export class Config {
   /**
    * Returns configuration, considering the stage. If no stage is set, just returns this.getConfigContent().
    */
-  get(): ProjectConfigFile {
-    const config: ProjectConfigFile = this.getContent();
+  get(): ProjectConfig {
+    const config: ProjectConfig = this.getContent();
 
     if (this.stage) {
-      const stagedConfig: ProjectConfigFile | undefined = _get(config, `stages.${this.stage}`);
+      const stagedConfig: ProjectConfig | undefined = _get(config, `stages.${this.stage}`);
       if (stagedConfig) {
         _mergeWith(config, stagedConfig, mergeArrayCustomizer);
 
@@ -69,9 +68,9 @@ export class Config {
   /**
    * Reads and returns the content of the project's config file.
    */
-  getContent(): ProjectConfigFile {
+  getContent(): ProjectConfig {
     try {
-      const config: ProjectConfigFile = require(this.getPath());
+      const config: ProjectConfig = require(this.getPath());
       return config;
     } catch (error) {
       throw new JovoCliError({
