+++ core/src/constants.ts
@@ -33,3 +33,4 @@ export const PACKAGE: string = SUPPORTS_EMOJI ? get('package') : '';
 export const DOWNLOAD: string = SUPPORTS_EMOJI ? get('inbox_tray') : '';
 export const MAGNIFYING_GLASS: string = SUPPORTS_EMOJI ? get('mag') : '';
 export const DISK: string = SUPPORTS_EMOJI ? get('floppy_disk') : '';
+export const CONSTRUCTION: string = SUPPORTS_EMOJI ? get('construction') : '';
